crewai-server/requirements.txt
crewai[tools]>=0.95.0
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
paramiko>=3.4.0
python-dotenv>=1.0.0
